+++ router/delivery/http/router_handler.go
@@ -137,7 +137,7 @@ func (a *RouterHandler) GetOptimalQuote(c echo.Context) (err error) {
 		scalingFactor = a.getSpotPriceScalingFactor(tokenIn.Denom, tokenOutDenom)
 	}
 
-	_, _, err = quote.PrepareResult(ctx, scalingFactor)
+	_, _, err = quote.PrepareResult(ctx, scalingFactor, a.logger)
 	if err != nil {
 		return err
 	}
@@ -186,7 +186,7 @@ func (a *RouterHandler) GetDirectCustomQuote(c echo.Context) error {
 		scalingFactor = a.getSpotPriceScalingFactor(tokenIn.Denom, tokenOutDenom[len(tokenOutDenom)-1])
 	}
 
-	_, _, err = quote.PrepareResult(ctx, scalingFactor)
+	_, _, err = quote.PrepareResult(ctx, scalingFactor, a.logger)
 	if err != nil {
 		return c.JSON(domain.GetStatusCode(err), domain.ResponseError{Message: err.Error()})
 	}
@@ -368,7 +368,8 @@ func getValidTokenInTokenOutStr(c echo.Context) (tokenOutStr, tokenInStr string,
 
 func getValidPoolID(c echo.Context) ([]uint64, error) {
 	// We accept two poolIDs and poolID parameters, and require at least one of them to be filled
-	poolIDStr := strings.Split(c.QueryParam("poolID"), ",")
+	poolIDParam := c.QueryParam("poolID")
+	poolIDStr := strings.Split(poolIDParam, ",")
 	if len(poolIDStr) == 0 {
 		return nil, errors.New("poolID is required")
 	}